Center labels around 0 for regression
- mf files: add 100 to all labels to avoid marsyas bug with
  negative labels
- read into python: subtract 100 to center around 0
- all display code: only deal with centered values

+++ src/vivi_controller.h
@@ -70,7 +70,7 @@ private:
     double m_K[NUM_STRINGS][NUM_DYNAMICS];
     double m_dampen[NUM_STRINGS][NUM_DYNAMICS];
 
-    int cats[CATS_MEAN_LENGTH]; // TODO: to test mean
+    double cats[CATS_MEAN_LENGTH]; // TODO: to test mean
     int cats_index;
 
     void note_setup_actions(NoteBeginning begin);
